DROTH-916 Code clean up
